Change UnitModel's package and rename joystickDeadband to deadband
